+++ examples/copyConfigForTestsOnly/config/plugins.js
@@ -1,9 +1,0 @@
-// Plugins to be loaded using require by Impress
-
-module.exports = [
-  'impress.log',
-  'impress.security',
-  'api.uglify',
-  'impress.events',
-  'impress.sse'
-];