+++ examples/copyContentToProjectFolder/applications/example/config/plugins.js
@@ -1,9 +1,0 @@
-// Plugins that should be mixedin into application and Client instances
-
-module.exports = [
-  'impress.events',
-  'impress.sse',
-  'impress.websocket',
-  //'impress.passport',
-  'impress.state',
-];