+++ examples/impressCloudController/instanceC1/applications/example/config/plugins.js
@@ -1,7 +1,0 @@
-// Plugins that should be mixedin into application and Client instances
-
-module.exports = [
-  'impress.events',
-  'impress.sse',
-  //'impress.passport'
-];